new grid: ControlGrid44_flow
single input:  a ControlGrid44. return a ControlGrid44 'type' object, which is a new version of the input where the inner control points now 'flow' instead of being limited by G1 edge connection rules. for each of the 4 inner control points, a parameter is available to force back an edge to G1 connection rules.
another parameter is in place, but not developed, that will allow maintaining  only 'mirrorability', while still improving flow, which is a lesser requirement than G1 connectivity.
So mirrorabilty is available, but it is less 'flowy' than what the future parameter will allow.