blend weight fix & SilkPose function
blend weight error checking was the largest source of errors. difficult to replicate the initial problem i wanted to address, so that part may or may not work. at least, i'm not overwritng good stuff.

SilkPose: a doc object that exist for the sole purpose of attaching other objects to it. this is a blend od MapMode's "ObjectXY/ZX/YX"  and "translate origin"